% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-histogram-rounded.R
\name{geom_histogram_rounded}
\alias{geom_histogram_rounded}
\title{Histogram with rounded bars}
\usage{
geom_histogram_rounded(
  mapping = NULL,
  data = NULL,
  position = "stack",
  radius = 0.2,
  ...,
  binwidth = NULL,
  bins = NULL,
  center = NULL,
  boundary = NULL,
  closed = c("right", "left"),
  pad = FALSE,
  breaks = NULL,
  drop = "none",
  orientation = NA,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{ggplot2::fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{radius}{A normalized rounding amount between 0 and 1. Use \code{0} for
square corners and \code{1} for the maximum rounding each bar can safely
support.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{ggplot2::layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{binwidth}{The width of the bins. The default uses 30 bins that cover
the range of the data. You should usually override this value, exploring
multiple widths to find the best to illustrate the stories in your data.}

\item{bins}{Number of bins. Overridden by \code{binwidth}. Defaults to 30.}

\item{center, boundary}{Bin position specifiers passed to
\code{\link[ggplot2:geom_histogram]{ggplot2::stat_bin()}}.}

\item{closed}{One of \code{"right"} or \code{"left"} indicating whether right or
left edges of bins are included.}

\item{pad}{If \code{TRUE}, adds empty bins at either side of the range.}

\item{breaks}{A numeric vector of bin boundaries.}

\item{drop}{Controls whether empty bins are dropped. Passed to
\code{\link[ggplot2:geom_histogram]{ggplot2::stat_bin()}}.}

\item{orientation}{The orientation of the layer. Passed to
\code{\link[ggplot2:geom_histogram]{ggplot2::stat_bin()}}.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{ggplot2::borders()}}.}
}
\value{
A ggplot object.
}
\description{
Create a histogram with rounded corners using
'ggplot2'. \code{geom_histogram_rounded()} is an extension of the
\code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}} function and uses the same binning
parameters as \code{\link[ggplot2:geom_histogram]{ggplot2::stat_bin()}} while drawing bars with rounded
corners.
}
\examples{
library(ggplot2)

ggplot(faithful, aes(waiting)) +
  geom_histogram_rounded()

ggplot(faithful, aes(waiting)) +
  geom_histogram_rounded(bins = 10, radius = 1)
}
